Re-organize file structure and rename raft() into raft_large
