+++ torchvision/models/optical_flow/__init__.py
@@ -1,1 +1,1 @@
-from ._raft.raft import RAFT, raft, raft_small
+from .raft import RAFT, raft_large, raft_small
